revise user input

+++ celler/ij_port.py
@@ -40,7 +40,8 @@ class IJPort:
             self.imp = imp = channels[0]
         imp.getProcessor().resetMinAndMax()
         self.pixels = ij.py.from_java(imp).to_numpy()
-        np.save(self._np_data_path, self.pixels)
+        if not os.path.exists(self._np_data_path):
+            np.save(self._np_data_path, self.pixels)
         ij.ui().show(self.imp)
         self.roi_manager = ij.RoiManager.getRoiManager()
         ij.RoiManager()
@@ -54,12 +55,10 @@ class IJPort:
         os.makedirs(self.log_folder, exist_ok=True)
         self.roi_manager = self.dataset = self.imp = self.ij = None
         self.pixels: Optional[np.ndarray] = None
+        self.rois = list()
 
-    def select_last_roi(self):
-        self.roi_manager.deselect()
-        count = self.roi_manager.getCount()
-        self.roi_manager.select(count-1)
-        return self.roi_manager.getRoi(count-1)
+    def retrieve_rois(self):
+        self.rois = [self.roi_manager.getRoi(i) for i in range(self.roi_manager.getCount())]
 
     def find_blobs(self, frame: int):
         img = self.pixels[frame]
@@ -71,28 +70,72 @@ class IJPort:
         cell_mask_clean = morphology.remove_small_holes(cell_mask_clean, self.config.max_hole)
         label_mask = measure.label(cell_mask_clean)
         region_properties = measure.regionprops(label_mask)
-        label_mask_clean = label_mask.copy()
         for r in region_properties:
             if r.area > self.config.max_size:
-                label_mask_clean[label_mask_clean == r.label] = 0
+                cell_mask_clean[label_mask == r.label] = 0
+        label_mask_clean = measure.label(cell_mask_clean)
         region_properties = measure.regionprops(label_mask_clean)
-        return smooth, label_mask_clean, region_properties, Mask(label_mask_clean).polygons().points
+        return smooth, label_mask_clean, region_properties
+
+    @staticmethod
+    def mask2polygon(mask):
+        return Mask(mask).polygons().points
+
+    def _plot(self, smooth, label_mask_clean, region_properties, output_name):
+        fig, ax = plt.subplots(1, 1, figsize=(20, 20))
+        ax.imshow(smooth)
+        ax.contour(label_mask_clean, colors='red')
+        for idx, rp in enumerate(region_properties):
+            ax.text(rp.centroid[1], rp.centroid[0], str(idx))
+        fig.tight_layout()
+        fig.savefig(os.path.join(self.log_folder, f'{output_name}.pdf'))
 
     def plot(self):
         if os.path.exists(self._np_data_path):
             self.pixels = np.load(self._np_data_path)
         else:
             self.init_ij()
-        smooth, label_mask_clean, region_properties, polygons = self.find_blobs(0)
-        fig, ax = plt.subplots(1, 1, figsize=(20, 20))
-        ax.imshow(smooth)
-        ax.contour(label_mask_clean, colors='red')
-        fig.tight_layout()
-        fig.savefig(os.path.join(self.log_folder, 'cells.pdf'))
+        smooth, label_mask_clean, region_properties = self.find_blobs(0)
+        self._plot(smooth, label_mask_clean, region_properties, 'cells')
 
     def segment(self):
         self.init_ij()
         logger.warning("Select your cell.")
-        # input('Press Enter to proceed.')
-        roi = self.select_last_roi()
-        x = 1
+        input('Press Enter to proceed.')
+        self.retrieve_rois()
+        self.find_closest(self.rois[-1], 0)
+        input('Please check the revision.')
+
+    @staticmethod
+    def read_roi(roi):
+        n = roi.getNCoordinates()
+        xs, ys = np.array(roi.getXCoordinates()[:n], dtype=float), np.array(roi.getYCoordinates()[:n], dtype=float)
+        xs += roi.getXBase()
+        ys += roi.getYBase()
+        coordinates = np.array([xs, ys]).T
+        center = np.array([roi.getBounds().getCenterX(), roi.getBounds().getCenterY()])
+        return coordinates, center
+
+    def find_closest(self, input_roi, frame: int):
+        input_coordinates, input_center = self.read_roi(input_roi)
+
+        distances = list()
+        smooth, label_mask_clean, region_properties = self.find_blobs(frame)
+        for rp in region_properties:
+            distances.append(np.sqrt(np.sum((np.array([rp.centroid[1], rp.centroid[0]]) - input_center)**2)))
+        cell_idx = np.argmin(distances)
+
+        polygon = self.mask2polygon(label_mask_clean == region_properties[cell_idx].label)[0].astype(float)
+        polygon_class = sj.jimport('ij.gui.PolygonRoi')
+        roi = polygon_class(polygon[:, 0].tolist(), polygon[:, 1].tolist(), polygon.shape[0], 2)
+        # overlay_class = sj.jimport('ij.gui.Overlay')
+        # ov = overlay_class()
+        # ov.add(roi)
+        self.roi_manager.addRoi(roi)
+        self.delete_roi(0)
+
+    def delete_roi(self, index: int):
+        self.retrieve_rois()
+        self.roi_manager.select(index)
+        self.roi_manager.runCommand('Delete')
+        self.retrieve_rois()
